+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.90] - 2021-06-17
+## [1.0.91] - 2021-06-21
+### Fixed
+- DO-101: Fields indent fixed
+- DO-101: Typo fixed
+
+## [1.0.90] - 2021-06-21
 ### Fixed
 - DO-101: "Invite Sent" labels could be located in one row with First and Last Names as it was before
 - DO-101: "Invitation message" label overlaps with a header while scrolling in mobile version